fix an issue related to minuit fit including gamma pdf which caused lots of failures
when generate an integer random number (in frequentist manner) from gamma pdf,  need +1 to it to be consistent with norminal one which is (N+1)

and change the default range of nuisances to be [-5, 5] instead of [-20, 20] which caused lots of more function calls.



CAVEAT:  when using ProfileLikelihood approximation method to estimate the significace s, and if s>4, then we need switch to [-20,20] to make sure the tail correct.