A.
Improvement in fits:   with option "--bMultiSigProcShareSamePDF"

in hzz4l and hzz2l2q, there are several signal processes, which all share the identical parametrized pdf.

With the obove option, the program will avoid duplications of pdf evaluation.


B.

fix a bug related with the flat param, when assiging nominal values of those parameters


C.

add option "--bForceSymmetryError" to make all asymmetry uncertainties in the datacards symmetric (and conservative) inside of code


D.

a small improvement on the ouput of fit resutls

before, there are some chances fits are sucessful, but the parabola error bars are not correctly returned, now in this case, return the error from fit instead of parabola  errors.